all database sensor tests: patterns

+++ dqops/src/integration-test/java/com/dqops/duckdb/sensors/column/patterns/DuckdbColumnPatternsTextNotMatchingRegexCountSensorParametersSpecIntegrationTest.java
@@ -74,7 +74,7 @@ public class DuckdbColumnPatternsTextNotMatchingRegexCountSensorParametersSpecIn
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(0.0, ValueConverter.toDouble(resultTable.column(0).get(0)));
+        Assertions.assertEquals(0L, ValueConverter.toLong(resultTable.column(0).get(0)));
     }
 
     @Test
